moved home bg components (starbox and sweeps)
